Moves all the methods from the WPS 2.0 standard to a dedicated class.
