+++ server/src/main/java/org/orbiswps/server/utils/WpsServerUtils.java
@@ -1,0 +1,46 @@
+package org.orbiswps.server.utils;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.xnap.commons.i18n.I18n;
+import org.xnap.commons.i18n.I18nFactory;
+
+import javax.xml.datatype.DatatypeConfigurationException;
+import javax.xml.datatype.DatatypeFactory;
+import javax.xml.datatype.Duration;
+import javax.xml.datatype.XMLGregorianCalendar;
+import java.util.Date;
+import java.util.GregorianCalendar;
+
+/**
+ * Utility methods for the WpsServer.
+ *
+ * @author Sylvain PALOMINOS
+ */
+public class WpsServerUtils {
+
+    /** Logger */
+    private static final Logger LOGGER = LoggerFactory.getLogger(WpsServerUtils.class);
+    /** I18N object */
+    private static final I18n I18N = I18nFactory.getI18n(WpsServerUtils.class);
+
+    /**
+     * Creates a XMLGregorianCalendar object which represent the date of now + durationInMillis.
+     * @param durationInMillis Duration in milliseconds to add to thenow date.
+     * @return A XMLGregorianCalendar object which represent the date of now + durationInMillis.
+     */
+    public static XMLGregorianCalendar getXMLGregorianCalendar(long durationInMillis){
+        GregorianCalendar calendar = new GregorianCalendar();
+        calendar.setTime(new Date());
+        XMLGregorianCalendar date = null;
+        try {
+            DatatypeFactory datatypeFactory = DatatypeFactory.newInstance();
+            date = datatypeFactory.newXMLGregorianCalendar(calendar);
+            Duration duration = datatypeFactory.newDuration(durationInMillis);
+            date.add(duration);
+        } catch (DatatypeConfigurationException e) {
+            LOGGER.error(I18N.tr("Unable to generate the XMLGregorianCalendar object.\nCause : {0}.", e.getMessage()));
+        }
+        return date;
+    }
+}
